requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.9.2
python-dotenv>=1.0.1
langchain-google-genai>=0.0.11
browser-use>=0.1.0
playwright>=1.42.0 